Moved common activity logic to the library module.
