+++ app/src/main/java/com/rohankhayech/choona/view/activity/TunerActivity.kt
@@ -18,20 +18,11 @@
 
 package com.rohankhayech.choona.view.activity
 
-import java.io.IOException
-import android.Manifest
 import android.content.Intent
-import android.net.Uri
 import android.os.Build
 import android.os.Bundle
-import android.provider.Settings.ACTION_APPLICATION_DETAILS_SETTINGS
-import android.view.WindowManager
-import androidx.activity.ComponentActivity
-import androidx.activity.OnBackPressedCallback
-import androidx.activity.addCallback
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.activity.viewModels
 import androidx.compose.material3.windowsizeclass.ExperimentalMaterial3WindowSizeClassApi
 import androidx.compose.material3.windowsizeclass.WindowHeightSizeClass
 import androidx.compose.material3.windowsizeclass.WindowWidthSizeClass
@@ -42,34 +33,19 @@ import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.datastore.preferences.core.emptyPreferences
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.lifecycleScope
 import com.rohankhayech.choona.BuildConfig
 import com.rohankhayech.choona.R
-import com.rohankhayech.choona.controller.midi.MidiController
 import com.rohankhayech.choona.controller.play.ReviewController
 import com.rohankhayech.choona.controller.play.ReviewControllerImpl
-import com.rohankhayech.choona.model.preferences.InitialTuningType
 import com.rohankhayech.choona.model.preferences.TunerPreferences
 import com.rohankhayech.choona.model.preferences.TunerPreferences.Companion.REVIEW_PROMPT_ATTEMPTS
-import com.rohankhayech.choona.model.preferences.tunerPreferenceDataStore
 import com.rohankhayech.choona.model.tuning.TuningEntry
-import com.rohankhayech.choona.view.PermissionHandler
 import com.rohankhayech.choona.view.screens.MainLayout
 import com.rohankhayech.choona.view.screens.TunerErrorScreen
 import com.rohankhayech.choona.view.screens.TunerPermissionScreen
 import com.rohankhayech.choona.view.theme.AppTheme
-import com.rohankhayech.choona.view.viewmodel.TunerViewModel
-import com.rohankhayech.music.Instrument
-import com.rohankhayech.music.Tuning
 import kotlinx.coroutines.delay
-import kotlinx.coroutines.flow.Flow
-import kotlinx.coroutines.flow.catch
-import kotlinx.coroutines.flow.firstOrNull
-import kotlinx.coroutines.flow.map
-import kotlinx.coroutines.launch
-import org.billthefarmer.mididriver.GeneralMidiConstants
 
 /**
  * Activity that allows the user to select a tuning and tune their guitar, displaying a comparison of played notes
@@ -77,25 +53,8 @@ import org.billthefarmer.mididriver.GeneralMidiConstants
  *
  * @author Rohan Khayech
  */
-class TunerActivity : ComponentActivity() {
+class TunerActivity : BaseTunerActivity() {
 
-    /** View model used to hold the current tuner state. */
-    private val vm: TunerViewModel by viewModels()
-
-    /** Handler used to check and request microphone permission. */
-    private lateinit var ph: PermissionHandler
-
-    /** MIDI controller used to play guitar notes. */
-    private lateinit var midi: MidiController
-
-    /** User preferences for the tuner. */
-    private lateinit var prefs: Flow<TunerPreferences>
-
-    /** Callback used to dismiss tuning selection screen when the back button is pressed. */
-    private lateinit var dismissTuningSelectorOnBack: OnBackPressedCallback
-
-    /** Callback used to dismiss configure tuning panel when the back button is pressed. */
-    private lateinit var dismissConfigurePanelOnBack: OnBackPressedCallback
 
     /** Google Play review controller. */
     private lateinit var reviewController: ReviewController
@@ -109,54 +68,6 @@ class TunerActivity : ComponentActivity() {
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.R) {
             enableEdgeToEdge()
-        }
-
-        // Setup preferences
-        prefs = tunerPreferenceDataStore.data
-            .catch { e -> if (e is IOException) emit(emptyPreferences()) else throw e }
-            .map(TunerPreferences::fromAndroidPreferences)
-
-        // Setup permission handler.
-        ph = PermissionHandler(this, Manifest.permission.RECORD_AUDIO)
-
-        // Setup MIDI controller for note playback.
-        midi = MidiController(vm.tuner.tuning.value.numStrings())
-
-        // Load tunings
-        lifecycleScope.launch {
-            val firstLoad = vm.tuningList.loadTunings(this@TunerActivity)
-
-            // Initialize edit mode and initial tuning state from preferences only on app load.
-            if (firstLoad) prefs.firstOrNull()?.let { preferences ->
-                vm.setEditMode(preferences.editModeDefault)
-
-                // Switch to initial tuning
-                when(preferences.initialTuning) {
-                    InitialTuningType.PINNED -> when (vm.tuningList.pinned.value) {
-                        is TuningEntry.InstrumentTuning -> setTuning(vm.tuningList.pinned.value.tuning!!)
-                        is TuningEntry.ChromaticTuning -> vm.tuner.setChromatic(true)
-                    }
-                    InitialTuningType.LAST_USED -> vm.tuningList.lastUsed.value?.let {
-                        when (it) {
-                            is TuningEntry.InstrumentTuning -> setTuning(it.tuning)
-                            is TuningEntry.ChromaticTuning -> vm.tuner.setChromatic(true)
-                        }
-                    }
-                }
-            }
-        }
-
-        // Setup custom back navigation.
-        dismissConfigurePanelOnBack = onBackPressedDispatcher.addCallback(this,
-            enabled = vm.configurePanelOpen.value,
-        ) {
-            dismissConfigurePanel()
-        }
-
-        dismissTuningSelectorOnBack = onBackPressedDispatcher.addCallback(this,
-            enabled = vm.tuningSelectorOpen.value
-        ) {
-            dismissTuningSelector()
         }
 
         reviewController = ReviewControllerImpl(this)
@@ -239,43 +150,21 @@ class TunerActivity : ComponentActivity() {
                         autoDetect = autoDetect,
                         chromatic = chromatic,
                         favTunings = favTunings,
-                        getCanonicalName = {
-                            vm.tuningList.run {
-                                this@MainLayout.getCanonicalName()
-                            }
-                        },
+                        getCanonicalName = vm.tuningList::getCanonicalName,
                         prefs = prefs,
                         tuningList = vm.tuningList,
                         tuningSelectorOpen = tuningSelectorOpen,
                         configurePanelOpen = configurePanelOpen,
-                        onSelectString = remember(prefs.enableStringSelectSound) {
-                            {
-                                vm.tuner.selectString(it)
-                                // Play sound on string selection.
-                                if (prefs.enableStringSelectSound) playStringSelectSound(it)
-                            }
-                        },
+                        onSelectString = ::selectString,
                         onSelectTuning = ::setTuning,
-                        onSelectChromatic = { vm.tuner.setChromatic(true) },
-                        onSelectNote = remember(prefs.enableStringSelectSound) {
-                            {
-                                vm.tuner.selectNote(it)
-                                // Play sound on string selection.
-                                if (prefs.enableStringSelectSound) playNoteSelectSound(it)
-                            }
-                        },
+                        onSelectChromatic = vm.tuner::setChromatic,
+                        onSelectNote = ::selectNote,
                         onTuneUpString = vm.tuner::tuneStringUp,
                         onTuneDownString = vm.tuner::tuneStringDown,
                         onTuneUpTuning = vm.tuner::tuneUp,
                         onTuneDownTuning = vm.tuner::tuneDown,
                         onAutoChanged = vm.tuner::setAutoDetect,
-                        onTuned = remember(prefs.enableInTuneSound) {
-                            {
-                                vm.tuner.setTuned()
-                                // Play sound when string tuned.
-                                if (prefs.enableInTuneSound) playInTuneSound()
-                            }
-                        },
+                        onTuned = ::setTuned,
                         onOpenTuningSelector = ::openTuningSelector,
                         onSettingsPressed = ::openSettings,
                         onConfigurePressed = ::openConfigurePanel,
@@ -300,197 +189,6 @@ class TunerActivity : ComponentActivity() {
                 }
             }
         }
-
-        // Keep the screen on while tuning.
-        window.addFlags(WindowManager.LayoutParams.FLAG_KEEP_SCREEN_ON)
-    }
-
-    /** Called when the activity resumes after being paused. */
-    override fun onResume() {
-        // Call superclass.
-        super.onResume()
-
-        // Start midi driver.
-        midi.start()
-
-        // Start the tuner if no panels are open.
-        if (!vm.tuningSelectorOpen.value && !vm.configurePanelOpen.value) {
-            try {
-                vm.tuner.start(ph)
-            } catch (_: Exception) {
-                // Catch and ignore, error will be displayed in the UI.
-            }
-        }
-    }
-
-    /** Called when the activity is paused but still visible. */
-    override fun onPause() {
-        // Stop the tuner.
-        vm.tuner.stop()
-
-        // Stop midi driver.
-        midi.stop()
-
-        // Call superclass.
-        super.onPause()
-    }
-
-    /** Called when the activity is no longer visible. */
-    override fun onStop() {
-        // Save tunings.
-        vm.tuningList.saveTunings(this)
-
-        // Call superclass
-        super.onStop()
-    }
-
-    /** Plays the string selection sound for the specified [string]. */
-    private fun playStringSelectSound(string: Int) {
-        lifecycleScope.launch {
-            midi.playNote(
-                string,
-                MidiController.noteIndexToMidi(vm.tuner.tuning.value.getString(string).rootNoteIndex),
-                150,
-                vm.tuner.tuning.value.instrument.midiInstrument
-            )
-        }
-    }
-
-    /** Plays the note selection sound for the specified [noteIndex]. */
-    private fun playNoteSelectSound(noteIndex: Int) {
-        lifecycleScope.launch {
-            midi.playNote(
-                0,
-                MidiController.noteIndexToMidi(noteIndex),
-                150,
-                Instrument.GUITAR.midiInstrument
-            )
-        }
-    }
-
-    /** Plays the in tune sound for the selected string. */
-    private fun playInTuneSound() {
-        val string = if (vm.tuner.chromatic.value) 0 else vm.tuner.selectedString.value
-        val noteIndex = if (vm.tuner.chromatic.value) vm.tuner.selectedNote.value
-            else vm.tuner.tuning.value.getString(string).rootNoteIndex
-        lifecycleScope.launch {
-            midi.playNote(
-                string,
-                MidiController.noteIndexToMidi(noteIndex) + 12,
-                50,
-                GeneralMidiConstants.MARIMBA
-            )
-        }
-    }
-
-    /**
-     * Opens the configure tuning panel, and stops the tuner.
-     */
-    private fun openConfigurePanel() {
-        dismissConfigurePanelOnBack.isEnabled = true
-        vm.openConfigurePanel()
-        vm.tuner.stop()
-    }
-
-    /**
-     * Opens the tuning selection screen, and stops the tuner.
-     */
-    private fun openTuningSelector() {
-        dismissTuningSelectorOnBack.isEnabled = true
-        vm.openTuningSelector()
-        vm.tuner.stop()
-    }
-
-    /**
-     * Dismisses the tuning selection screen and restarts the tuner if no other panel is open.
-     */
-    private fun dismissTuningSelector() {
-        dismissTuningSelectorOnBack.isEnabled = false
-        vm.dismissTuningSelector()
-        if (!vm.configurePanelOpen.value) {
-            try {
-                vm.tuner.start(ph)
-            } catch(_: Exception) {}
-        }
-    }
-
-    /** Dismisses the configure panel and restarts the tuner if no other panel is open. */
-    private fun dismissConfigurePanel() {
-        dismissConfigurePanelOnBack.isEnabled = false
-        vm.dismissConfigurePanel()
-        if (!vm.tuningSelectorOpen.value) {
-            try {
-                vm.tuner.start(ph)
-            } catch (_: Exception) {}
-        }
-    }
-
-    /**
-     * Sets the current tuning to the [tuning] selected on the tuning
-     * selection screen, restarts the tuner if no other panel is open,
-     * and recreates the MIDI driver if necessary.
-     */
-    private fun selectTuning(tuning: Tuning) {
-        // Consume back stack entry.
-        dismissTuningSelectorOnBack.isEnabled = false
-
-        // Recreate MIDI driver if number of strings different.
-        checkAndRecreateMidiDriver(tuning)
-
-        // Select the tuning.
-        vm.selectTuning(tuning)
-
-        // Start tuner if no other panel is open.
-        if (!vm.configurePanelOpen.value) {
-            try {
-                vm.tuner.start(ph)
-            } catch(_: Exception) {}
-        }
-    }
-
-    /**
-     * Sets chromatic mode on as selected on the tuning selection screen
-     * and restarts the tuner if no other panel is open.
-     */
-    private fun selectChromatic() {
-        // Consume back stack entry.
-        dismissTuningSelectorOnBack.isEnabled = false
-
-        // Select the tuning.
-        vm.selectChromatic()
-
-        // Start tuner if no other panel is open.
-        if (!vm.configurePanelOpen.value) {
-            try {
-                vm.tuner.start(ph)
-            } catch(_: Exception) {}
-        }
-    }
-
-    /**
-     * Sets the current tuning to the [tuning] specified,
-     * and recreates the MIDI driver if necessary.
-     */
-    private fun setTuning(tuning: Tuning) {
-        // Recreate MIDI driver if number of strings different.
-        checkAndRecreateMidiDriver(tuning)
-
-        // Select the tuning.
-        vm.tuner.setTuning(tuning)
-    }
-
-    /**
-     * Recreates the MIDI driver when the number of strings
-     * in the new tuning is different from the current tuning.
-     *
-     * @param newTuning The new selected tuning.
-     */
-    private fun checkAndRecreateMidiDriver(newTuning: Tuning) {
-        if (newTuning.numStrings() != vm.tuner.tuning.value.numStrings()) {
-            midi.stop()
-            midi = MidiController(newTuning.numStrings())
-            midi.start()
-        }
     }
 
     /** Opens the tuner settings activity. */
@@ -503,16 +201,6 @@ class TunerActivity : ComponentActivity() {
         val intent = Intent(this, SettingsActivity::class.java)
         intent.putExtra(SettingsActivity.EXTRA_PINNED, pinnedName)
         startActivity(intent)
-    }
-
-    /** Opens the permission settings screen in the device settings. */
-    private fun openPermissionSettings() {
-        startActivity(
-            Intent(
-                ACTION_APPLICATION_DETAILS_SETTINGS,
-                Uri.fromParts("package", packageName, null)
-            )
-        )
     }
 
     /** Launches a prompt for the user to review the app on Google Play. */
