+++ app/src/main/java/com/rohankhayech/choona/view/components/TuningSelector.kt
@@ -244,7 +244,7 @@ fun TuningItem(
     tuning: TuningEntry,
     fontWeight: FontWeight,
     horizontalAlignment: Alignment.Horizontal = Alignment.Start,
-    getCanonicalName: TuningEntry.InstrumentTuning.() -> String,
+    getCanonicalName: (TuningEntry.InstrumentTuning) -> String,
 ) {
     val tuningName = when (tuning) {
         is TuningEntry.ChromaticTuning -> stringResource(R.string.chromatic)
@@ -252,7 +252,7 @@ fun TuningItem(
             if (tuning.tuning.hasName()) {
                 tuning.tuning.name
             } else {
-                tuning.getCanonicalName()
+                getCanonicalName(tuning)
             }
     }
 
@@ -298,7 +298,10 @@ private fun Preview() {
         TuningSelector(
             Modifier.padding(8.dp),
             tuning = TuningEntry.InstrumentTuning(Tuning.STANDARD),
-            favTunings = remember { mutableStateOf(setOf(TuningEntry.InstrumentTuning(Tuning.STANDARD), TuningEntry.InstrumentTuning(Tuning.DROP_D))) },
+            favTunings = remember { mutableStateOf(setOf(
+                TuningEntry.InstrumentTuning(Tuning.STANDARD),
+                TuningEntry.InstrumentTuning(Tuning.DROP_D)
+            )) },
             openDirect = false,
             onSelect = {},
             onTuneDown = {},
@@ -318,7 +321,10 @@ private fun EditOffPreview() {
         TuningSelector(
             Modifier.padding(8.dp),
             tuning = TuningEntry.InstrumentTuning(Tuning.STANDARD),
-            favTunings = remember { mutableStateOf(setOf(TuningEntry.InstrumentTuning(Tuning.STANDARD), TuningEntry.InstrumentTuning(Tuning.DROP_D))) },
+            favTunings = remember { mutableStateOf(setOf(
+                TuningEntry.InstrumentTuning(Tuning.STANDARD),
+                TuningEntry.InstrumentTuning(Tuning.DROP_D)
+            )) },
             getCanonicalName = { this.tuning.toString() },
             openDirect = false,
             onSelect = {},
@@ -339,7 +345,7 @@ private fun TuningItemPreview() {
         TuningItem(
             Modifier.padding(8.dp),
             tuning = TuningEntry.InstrumentTuning(Tuning.STANDARD),
-            getCanonicalName = { this.tuning.toString() },
+            getCanonicalName = { it.toString() },
             fontWeight = FontWeight.Bold
         )
     }
