+++ app/src/main/java/com/rohankhayech/choona/view/screens/ConfigureTuningScreen.kt
@@ -91,7 +91,7 @@ fun ConfigureTuningScreen(
     chromatic: Boolean,
     selectedNote: Int,
     favTunings: State<Set<TuningEntry>>,
-    getCanonicalName: TuningEntry.InstrumentTuning.() -> String,
+    getCanonicalName: (TuningEntry.InstrumentTuning) -> String,
     onTuneUpString: (Int) -> Unit,
     onTuneDownString: (Int) -> Unit,
     onTuneUpTuning: () -> Unit,
@@ -191,7 +191,7 @@ private fun Preview() {
             chromatic = false,
             selectedNote = -29,
             favTunings = remember { mutableStateOf(emptySet()) },
-            getCanonicalName = { this.tuning.toString() },
+            getCanonicalName = { it.toString() },
             onTuneUpString = {},
             onTuneDownString = {},
             onTuneUpTuning = {},
