+++ lib/src/main/AndroidManifest.xml
@@ -18,5 +18,4 @@
   -->
 
 <manifest>
-
 </manifest>
